requirements.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
audioread==3.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
customtkinter==5.1.2
cycler==0.11.0
darkdetect==0.8.0
decorator==5.1.1
et-xmlfile==1.1.0
ffmpeg==1.4
Flask==2.2.3
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.51.1
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pooch==1.6.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-version==0.0.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
soundfile==0.11.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.14.1
